Also delete from documents

+++ application/routes.py
@@ -557,13 +557,26 @@ def get_registered_search_forms(request_id):
 def delete_all_reg_forms(date, reg_no):
     cursor = connect(cursor_factory=psycopg2.extras.DictCursor)
     try:
+        cursor.execute('select document_id from registered_documents '
+                       'where number=%(no)s and date=%(date)s', {
+                           'no': reg_no, 'date': date
+                       })
+        rows = cursor.fetchall()
+        if len(rows) != 1:
+            raise CaseworkAPIError("Could not retrieve unique document id")
+
+        document_id = rows[0]['document_id']
         cursor.execute('delete from registered_documents '
                        'where number=%(no)s and date=%(date)s', {
                            'no': reg_no, 'date': date
                        })
+
+        cursor.execute('delete from documents where document_id = %(id)s', {
+            'id': document_id
+        })
         return Response(status=200)
 
-    # TODO: also remove form from documents table?
+    # TODO: also remove form from documents table? <<- TEST! IMPORTANT!
     finally:
         complete(cursor)
